// src/components/Typo/types.tsx
export enum TypoColors {
  inherit = 'inherit',
  black = 'black',
  greyDark = 'greyDark',
  greyMedium = 'greyMedium',
  white = 'white',
  blue = 'blue',
  red = 'red',
}

export enum TypoVariants {
  h1 = 'h1',
  h2 = 'h2',
  h3 = 'h3',
  h4 = 'h4',
  h5 = 'h5',
  body1 = 'body1',
  body2 = 'body2',
  button = 'button',
  subTitle = 'subTitle',
}

export enum TypoAlignment {
  left = 'left',
  middle = 'middle',
}